refactor: create ordertable package
